@@ -13,24 +13,25 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/common": "~10.0.0",
-    "@angular/core": "~10.0.0",
-    "@angular/forms": "~10.0.0",
-    "@angular/platform-browser": "~10.0.0",
-    "@angular/platform-browser-dynamic": "~10.0.0",
-    "@angular/router": "~10.0.0",
-    "@ionic-native/camera": "^5.28.0",
-    "@ionic-native/core": "^5.0.0",
-    "@ionic-native/file": "^5.28.0",
-    "@ionic-native/file-path": "^5.28.0",
-    "@ionic-native/ionic-webview": "^5.28.0",
-    "@ionic-native/splash-screen": "^5.0.0",
-    "@ionic-native/status-bar": "^5.0.0",
-    "@ionic/angular": "^5.0.0",
-    "rxjs": "~6.5.5",
-    "speak-tts": "^2.0.8",
-    "tslib": "^2.0.0",
-    "zone.js": "~0.10.3"
+    "@angular/cli": "^11.0.1",
+    "@angular/common": "^11.0.0",
+    "@angular/core": "^11.0.0",
+    "@angular/forms": "^11.0.0",
+    "@angular/platform-browser": "^11.0.0",
+    "@angular/platform-browser-dynamic": "^11.0.0",
+    "@capacitor/core": "2.4.2",
+    "@ionic-native/camera": "^5.29.0",
+    "@ionic-native/core": "^5.29.0",
+    "@ionic-native/splash-screen": "^5.29.0",
+    "@ionic-native/status-bar": "^5.29.0",
+    "@ionic/angular": "^5.4.4",
+    "@ionic/angular-toolkit": "^2.3.3",
+    "@ionic/cli": "^6.12.1",
+    "cordova-plugin-camera": "^5.0.1",
+    "cordova-res": "^0.15.2",
+    "native-run": "^1.2.2",
+    "rxjs": "^6.6.3",
+    "speak-tts": "^2.0.8"
   },
   "devDependencies": {
     "@angular-devkit/build-angular": "^0.1100.1",
