@@ -15,14 +15,19 @@ import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { LoginComponent } from './pages/login/login.component';
 import { RegisterComponent } from './pages/register/register.component';
 import { AddCollectionComponent } from './pages/addcollection/addcollection.component';
+import { PageService } from './page.service';
+import { CommonModule } from '@angular/common';
+import { ItemComponent } from './pages/item/item.component';
+import { AddItemComponent } from './pages/add-item/add-item.component';
 
 @NgModule({
-  declarations: [AppComponent, LoginComponent, RegisterComponent, AddCollectionComponent],
+  declarations: [AppComponent, LoginComponent, RegisterComponent, AddCollectionComponent, ItemComponent, AddItemComponent],
   entryComponents: [],
-  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, HttpClientModule, FormsModule, ReactiveFormsModule],
+  imports: [BrowserModule, IonicModule.forRoot(), AppRoutingModule, HttpClientModule, FormsModule, ReactiveFormsModule, CommonModule],
   providers: [
     StatusBar,
     SplashScreen,
+    PageService,
     { provide: RouteReuseStrategy, useClass: IonicRouteStrategy },
     { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true }
   ],
