@@ -1,6 +1,8 @@
 import { NgModule } from '@angular/core';
 import { PreloadAllModules, RouterModule, Routes } from '@angular/router';
+import { AddItemComponent } from './pages/add-item/add-item.component';
 import { AddCollectionComponent as AddCollectionComponent } from './pages/addcollection/addcollection.component';
+import { ItemComponent } from './pages/item/item.component';
 import { LoginComponent } from './pages/login/login.component';
 import { RegisterComponent } from './pages/register/register.component';
 
@@ -11,7 +13,9 @@ const routes: Routes = [
   },
   { path: "addcollection", component: AddCollectionComponent },
   { path: "login", component: LoginComponent },
-  { path: "register", component: RegisterComponent }
+  { path: "register", component: RegisterComponent },
+  { path: "item", component: ItemComponent },
+  { path: "additem", component: AddItemComponent }
 ];
 @NgModule({
   imports: [
